login funcionando sem session

+++ src/main/java/Controle/Helpers/MenuLivroFuncionarioHelper.java
@@ -41,7 +41,7 @@ public class MenuLivroFuncionarioHelper {
                livro.getAutor(),
                livro.getAnoPubli(),
                (livro.getNumAvaliacoes() > 0) 
-                ? livro.getTotalAvaliacao() / livro.getNumAvaliacoes()
+                ? (double)livro.getTotalAvaliacao() / (double)livro.getNumAvaliacoes()
                 : "Sem avaliações",
                livro.getStatus()
             });
